Doofus_Game
Doofus is a simple yet engaging game that puts your reflexes and timing to the test. Created as one of my first Unity projects, this game is all about guiding a cube across platforms that randomly spawn and disappear after a short time. Your goal is to move swiftly to the next platform before the timer runs out, or risk watching the platform crumble beneath you, ending the game. With its fast-paced mechanics and challenging gameplay, Survive the Platforms offers a fun experience while showcasing the foundations of my game development journey.